# api-docs.yaml
openapi: 3.0.0
paths:
...
  /api/v1/products:
    get:
      operationId: ProductsController_findAll
      summary: Get all products with filters
      parameters: []
      responses:
        '200':
          description: Products retrieved successfully
      tags: &ref_6
        - products
    post:
      operationId: ProductsController_create
      summary: Create a new product
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDto'
      responses:
        '201':
          description: Product created successfully
        '400':
          description: Invalid input
      tags: *ref_6
      security:
        - bearer: []
  /api/v1/products/{id}:
    get:
      operationId: ProductsController_findById
      summary: Get product by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Product retrieved successfully
        '404':
          description: Product not found
      tags: *ref_6
    put:
      operationId: ProductsController_update
      summary: Update a product
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductDto'
      responses:
        '200':
          description: Product updated successfully
        '403':
          description: Forbidden - not the owner
        '404':
          description: Product not found
      tags: *ref_6
      security:
        - bearer: []
    delete:
      operationId: ProductsController_delete
      summary: Delete a product (soft delete)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully
        '403':
          description: Forbidden - not the owner
        '404':
          description: Product not found
      tags: *ref_6
      security:
        - bearer: []
...
  /api/v1/reservations:
    post:
      operationId: ReservationsController_create
      summary: Create a new reservation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationDto'
      responses:
        '201':
          description: Reservation created successfully
        '400':
          description: Invalid request
      tags: &ref_11
        - reservations
      security: &ref_12
        - bearer: []
  /api/v1/reservations/my:
    get:
      operationId: ReservationsController_getMyReservations
      summary: Get current user reservations
      parameters: []
      responses:
        '200':
          description: Reservations retrieved successfully
      tags: *ref_11
      security: *ref_12
  /api/v1/reservations/my-sales:
    get:
      operationId: ReservationsController_getMySales
      summary: Get reservations for current user products
      parameters: []
      responses:
        '200':
          description: Reservations retrieved successfully
      tags: *ref_11
      security: *ref_12
  /api/v1/reservations/product/{productId}:
    get:
      operationId: ReservationsController_getByProductId
      summary: Get reservations for a product
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Reservations retrieved successfully
      tags: *ref_11
      security: *ref_12
  /api/v1/reservations/{id}:
    get:
      operationId: ReservationsController_findById
      summary: Get reservation by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Reservation retrieved successfully
        '404':
          description: Reservation not found
      tags: *ref_11
      security: *ref_12
    put:
      operationId: ReservationsController_update
      summary: Update reservation status
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReservationDto'
      responses:
        '200':
          description: Reservation updated successfully
        '403':
          description: Forbidden
        '404':
          description: Reservation not found
      tags: *ref_11
      security: *ref_12
  /api/v1/reservations/test:
    get:
      operationId: ReservationsController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_11
      security: *ref_12
  /api/v1/reports:
    post:
      operationId: ReportsController_create
      summary: Create a new report
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportDto'
      responses:
        '201':
          description: Report created successfully
        '400':
          description: Invalid request
      tags: &ref_13
        - reports
      security: &ref_14
        - bearer: []
    get:
      operationId: ReportsController_findAll
      summary: Get all reports (Admin only)
      parameters:
        - name: status
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Reports retrieved successfully
      tags: *ref_13
      security: *ref_14
  /api/v1/reports/{id}:
    get:
      operationId: ReportsController_findById
      summary: Get report by ID (Admin only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Report retrieved successfully
        '404':
          description: Report not found
      tags: *ref_13
      security: *ref_14
  /api/v1/reports/{id}/status:
    put:
      operationId: ReportsController_updateStatus
      summary: Update report status (Admin only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Report status updated successfully
        '403':
          description: Forbidden
      tags: *ref_13
      security: *ref_14
  /api/v1/reports/test:
    get:
      operationId: ReportsController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_13
      security: *ref_14
...
info:
  title: memi.az API
  description: Recommerce Platform Backend API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateProductDto:
      type: object
      properties: {}
    UpdateProductDto:
      type: object
      properties: {}
...
    CreateReservationDto:
      type: object
      properties: {}
    UpdateReservationDto:
      type: object
      properties: {}
    CreateReportDto:
      type: object
      properties: {}
...