# src/modules/reservations/dtos/create-reservation.dto.ts
import { IsUUID } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class CreateReservationDto {
  @ApiProperty({
    example: '123e4567-e89b-12d3-a456-426614174000',
    description: 'Product ID to reserve',
  })
  @IsUUID()
  productId: string;
}